+++ src/main/kotlin/net/plpgsql/ideadebugger/PlFunctionSource.kt
@@ -9,11 +9,12 @@ import com.intellij.openapi.project.Project
 import com.intellij.openapi.vfs.VirtualFileSystem
 import com.intellij.psi.PsiDocumentManager
 import com.intellij.psi.PsiManager
+import com.intellij.psi.util.PsiElementFilter
 import com.intellij.psi.util.PsiTreeUtil
 import com.intellij.sql.psi.SqlBlockStatement
+import com.intellij.sql.psi.impl.SqlTokenElement
 import com.intellij.testFramework.LightVirtualFile
 import java.nio.charset.Charset
-import kotlin.properties.Delegates
 
 class PlFunctionSource(project: Project, def: PlApiFunctionDef) : LightVirtualFile(
     "${def.schema}.${def.name}",
@@ -25,6 +26,10 @@ class PlFunctionSource(project: Project, def: PlApiFunctionDef) : LightVirtualFi
     val md5 = def.md5
     var start: Int = 0
     var end: Int = 0
+    var codeRange = Pair(0, 0)
+    val lineRangeCount: Int by lazy {
+        codeRange.second - codeRange.first
+    }
 
     init {
         runReadAction {
@@ -33,11 +38,22 @@ class PlFunctionSource(project: Project, def: PlApiFunctionDef) : LightVirtualFi
                     PsiDocumentManager.getInstance(psi.project).getDocument(psi)?.let { doc ->
                         start = doc.getLineNumber(block.textOffset) - 2
                         end = doc.getLineNumber(block.textOffset + block.textLength) - 2
+
+                        val beginEl = PsiTreeUtil.collectElements(block, PsiElementFilter {
+                            it is SqlTokenElement && it.text.uppercase() == "BEGIN"
+                        }).firstOrNull()
+
+                        val endEl = PsiTreeUtil.collectElements(block, PsiElementFilter {
+                            it is SqlTokenElement && it.text.uppercase() == "END"
+                        }).lastOrNull()
+
+                        if (beginEl != null && endEl != null) {
+                            codeRange = Pair(doc.getLineNumber(beginEl.textOffset), doc.getLineNumber(endEl.textOffset))
+                        }
                     }
                 }
             }
         }
-
     }
 
     override fun getCharset(): Charset = Charsets.UTF_8
